msm: pm: fix power driver compilation issues
Fix power driver compilation issues. Modify basic data types
usage(i.e long & pointer sizes) such that make power drivers
more generic to compile for different build environments.

CRs-fixed: 579034

+++ arch/arm/mach-msm/msm-pm.c
@@ -1093,7 +1093,7 @@ static int msm_pc_debug_counters_copy(
 	return data->len;
 }
 
-static int msm_pc_debug_counters_file_read(struct file *file,
+static ssize_t msm_pc_debug_counters_file_read(struct file *file,
 		char __user *bufu, size_t count, loff_t *ppos)
 {
 	struct msm_pc_debug_counters_buffer *data;
@@ -1131,7 +1131,7 @@ static int msm_pc_debug_counters_file_open(struct inode *inode,
 		sizeof(struct msm_pc_debug_counters_buffer), GFP_KERNEL);
 
 	if (!file->private_data) {
-		pr_err("%s: ERROR kmalloc failed to allocate %d bytes\n",
+		pr_err("%s: ERROR kmalloc failed to allocate %zu bytes\n",
 		__func__, sizeof(struct msm_pc_debug_counters_buffer));
 
 		return -ENOMEM;
